requirements.txt
pandas
flask
keras
pillow
requests
google-cloud-storage
tensorflow==1.15